+++ vhost-device-media/CHANGELOG.md
@@ -5,7 +5,7 @@
 
 ### Changed
 
-- [[#987]](https://github.com/rust-vmm/vhost-device/pull/987) Update v4l2r dependency to v0.0.8.
+- [[#987]](https://github.com/rust-vmm/vhost-device/pull/987) Update v4l2r and virtio-media dependencies to v0.0.8 and removed ffmpeg feature.
 
 ### Fixed